Make the AOSP Merge script a bit smarter You no longer have to specify which repos to update.
The new script will look for any repos that Google has, and we track
via the manifest.

It also keeps track of which upstream merge's failed, and
tells the user when the script is done.

Another improvement to @Mazda-- 's script is that is only
removes the repos you need and *should* be faster, as well
that it is more portable, the only thing other ROMs would
need to change is the WORKING_DIR.

Change-Id: Ic71be398166d6497258290a223c9f0e899208c41
Signed-off-by: Jacob McSwain <jacob.a.mcswain@gmail.com>